refactor: move autoCheckin services to nested checkin directory structure
Move all autoCheckin-related services into a nested directory structure
under `services/checkin/autoCheckin/` to better organize the codebase
and group related functionality together.
