+++ entrypoints/background/runtimeMessages.ts
@@ -7,8 +7,8 @@ import {
 import { applyActionClickBehavior } from "~/entrypoints/background/actionClickBehavior"
 import { handleAccountKeyRepairMessage } from "~/services/accounts/accountKeyAutoProvisioning"
 import { handleAutoRefreshMessage } from "~/services/accounts/autoRefreshService"
-import { handleAutoCheckinMessage } from "~/services/autoCheckin/scheduler"
 import { handleChannelConfigMessage } from "~/services/channelConfigStorage"
+import { handleAutoCheckinMessage } from "~/services/checkin/autoCheckin/scheduler"
 import { handleExternalCheckInMessage } from "~/services/checkin/externalCheckInService"
 import { handleDailyBalanceHistoryMessage } from "~/services/dailyBalanceHistory/scheduler"
 import { handleLdohSiteLookupMessage } from "~/services/integrations/ldohSiteLookup/background"
