+++ services/checkin/autoCheckin/providers/shared.ts
@@ -5,7 +5,7 @@
  * magic strings (message keys, message parsing heuristics) across backends.
  */
 
-import type { AutoCheckinProviderResult } from "~/services/autoCheckin/providers/types"
+import type { AutoCheckinProviderResult } from "~/services/checkin/autoCheckin/providers/types"
 import { CHECKIN_RESULT_STATUS } from "~/types/autoCheckin"
 
 export const AUTO_CHECKIN_PROVIDER_FALLBACK_MESSAGE_KEYS = {
