+++ services/checkin/autoCheckin/providers/types.ts
